Refactored fragment to display the user list to the admin.
- Just to stay consistent with the conventions set by the service provider fragments.
