+++ app/src/main/java/com/arom/jobzi/fragment/AdminFragment.java
@@ -14,6 +14,7 @@ import com.arom.jobzi.R;
 import com.arom.jobzi.ServiceEditorActivity;
 import com.arom.jobzi.dapater.CustomPagerAdapter;
 import com.arom.jobzi.fragment.admin.AdminServicesFragment;
+import com.arom.jobzi.fragment.admin.AdminUsersFragment;
 import com.arom.jobzi.profile.ServiceProviderProfile;
 import com.arom.jobzi.service.Service;
 import com.arom.jobzi.util.Util;
@@ -43,10 +44,10 @@ public class AdminFragment extends Fragment implements AdminServicesFragment.Ser
 		
 		CustomPagerAdapter customPagerAdapter = new CustomPagerAdapter(getActivity().getSupportFragmentManager());
 		
-		UserListFragment userListFragment = new UserListFragment();
+		AdminUsersFragment adminUsersFragment = new AdminUsersFragment();
 		AdminServicesFragment adminServicesFragment = AdminServicesFragment.newInstance(this);
 		
-		customPagerAdapter.addFragment(userListFragment, getText(R.string.users_label));
+		customPagerAdapter.addFragment(adminUsersFragment, getText(R.string.users_label));
 		customPagerAdapter.addFragment(adminServicesFragment, getText(R.string.services_label));
 		
 		viewPager.setAdapter(customPagerAdapter);
